refactor(node): split dispatcher handlers
Extract control, announcement, data, insert, probe, offer, swap, and not-routable handling into dedicated classes.
Move routed message tracking into NodeRoutedMessageRouter and add swap handler tests.
